[SH-19699] [Android] Icon Label AnnotatedString 호환성 복구
