+++ sdg/src/main/java/com/shopl/sdg/component/icon_label/SDGIconLabel.kt
@@ -10,6 +10,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
@@ -121,6 +122,55 @@ fun SDGIconLabel(
     isFillMaxWidth: Boolean = true,
 ) {
     SDGIconLabel(
+        label = text,
+        labelColor = textColor,
+        size = size,
+        fontWeight = type.fontWeight,
+        gap = spacing.gap,
+        onClick = onClickLeftIcon ?: onClickRightIcon,
+        labelOverflow = SDGIconLabelOverflow.fromMaxLines(maxLines),
+        modifier = modifier,
+        leftIc = leftIconResId?.let { resId ->
+            SDGIconLabelIcon(
+                resId = resId,
+                tint = leftIconTint,
+            )
+        },
+        rightIc = rightIconResId?.let { resId ->
+            SDGIconLabelIcon(
+                resId = resId,
+                tint = rightIconTint,
+            )
+        },
+        isFillMaxWidth = isFillMaxWidth,
+    )
+}
+
+/**
+ * 신규 Icon Label API와의 하위 호환성을 위한 레거시 API입니다.
+ */
+@Deprecated(
+    message = "text/type/spacing/maxLines 대신 label/fontWeight/gap/labelOverflow를 사용하세요.",
+)
+@Composable
+@Suppress("DEPRECATION")
+fun SDGIconLabel(
+    text: AnnotatedString,
+    textColor: Color,
+    size: SDGIconLabelSize,
+    type: SDGIconLabelType,
+    spacing: SDGIconLabelSpacing,
+    modifier: Modifier = Modifier,
+    @DrawableRes leftIconResId: Int? = null,
+    leftIconTint: Color? = null,
+    onClickLeftIcon: (() -> Unit)? = null,
+    @DrawableRes rightIconResId: Int? = null,
+    rightIconTint: Color? = null,
+    onClickRightIcon: (() -> Unit)? = null,
+    maxLines: Int = Int.MAX_VALUE,
+    isFillMaxWidth: Boolean = true,
+) {
+    com.shopl.sdg.component.util.icon_label.SDGIconLabel(
         label = text,
         labelColor = textColor,
         size = size,
