+++ sdg/src/main/java/com/shopl/sdg/component/util/icon_label/SDGIconLabel.kt
@@ -1,6 +1,5 @@
 package com.shopl.sdg.component.util.icon_label
 
-import androidx.annotation.DrawableRes
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -17,8 +16,6 @@ import com.shopl.sdg.component.icon_label.SDGIconLabelGap
 import com.shopl.sdg.component.icon_label.SDGIconLabelIcon
 import com.shopl.sdg.component.icon_label.SDGIconLabelOverflow
 import com.shopl.sdg.component.icon_label.SDGIconLabelSize
-import com.shopl.sdg.component.icon_label.SDGIconLabelSpacing
-import com.shopl.sdg.component.icon_label.SDGIconLabelType
 import com.shopl.sdg_common.ext.clickable
 import com.shopl.sdg_common.ui.components.SDGImage
 import com.shopl.sdg_common.ui.components.SDGText
@@ -92,55 +89,6 @@ fun SDGIconLabel(
             )
         }
     }
-}
-
-/**
- * 신규 Icon Label API와의 하위 호환성을 위한 레거시 API입니다.
- */
-@Deprecated(
-    message = "text/type/spacing/maxLines 대신 label/fontWeight/gap/labelOverflow를 사용하세요.",
-)
-@Composable
-@Suppress("DEPRECATION")
-fun SDGIconLabel(
-    text: AnnotatedString,
-    textColor: Color,
-    size: SDGIconLabelSize,
-    type: SDGIconLabelType,
-    spacing: SDGIconLabelSpacing,
-    modifier: Modifier = Modifier,
-    @DrawableRes leftIconResId: Int? = null,
-    leftIconTint: Color? = null,
-    onClickLeftIcon: (() -> Unit)? = null,
-    @DrawableRes rightIconResId: Int? = null,
-    rightIconTint: Color? = null,
-    onClickRightIcon: (() -> Unit)? = null,
-    maxLines: Int = Int.MAX_VALUE,
-    isFillMaxWidth: Boolean = true,
-) {
-    SDGIconLabel(
-        label = text,
-        labelColor = textColor,
-        size = size,
-        fontWeight = type.fontWeight,
-        gap = spacing.gap,
-        onClick = onClickLeftIcon ?: onClickRightIcon,
-        labelOverflow = SDGIconLabelOverflow.fromMaxLines(maxLines),
-        modifier = modifier,
-        leftIc = leftIconResId?.let { resId ->
-            SDGIconLabelIcon(
-                resId = resId,
-                tint = leftIconTint,
-            )
-        },
-        rightIc = rightIconResId?.let { resId ->
-            SDGIconLabelIcon(
-                resId = resId,
-                tint = rightIconTint,
-            )
-        },
-        isFillMaxWidth = isFillMaxWidth,
-    )
 }
 
 private val SDGIconLabelIconSize = 14.dp
